refactor(replays): adds useCrumbHandlers util
So we're not repeating the same piece of code in multiple places
